Dev uint64

+++ libraries/protocol/include/graphene/protocol/asset_ops.hpp
@@ -398,7 +398,7 @@ namespace graphene { namespace protocol {
    {
       struct fee_parameters_type { 
          uint64_t fee = 20 * GRAPHENE_BLOCKCHAIN_PRECISION; 
-         uint32_t price_per_kbyte = GRAPHENE_BLOCKCHAIN_PRECISION;
+         uint64_t price_per_kbyte = GRAPHENE_BLOCKCHAIN_PRECISION;
       };
 
       asset            fee;
